Update to resource images and readme file.
All resource images should now have the proper font hinting for the section titles.

The README.txt file now warns against using  game art with this theme, and advises the use of the latest version of Simplemenu.

The theme.ini file has been updated to add features of Simplemenu 9.5, which allows game art to be disabled.